% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_functions.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Cleans NOAA earthquake data}
\usage{
eq_clean_data(raw_df)
}
\arguments{
\item{raw_df}{The raw dataframe created from data in the NOAA file.}
}
\value{
This function returns the clean earthquake data in a tbl_df object.
}
\description{
This is a function which generates a DATE column and changes LATITUDE, LONGITUDE,
EQ_PRIMARY, and DEATHS to numeric variables.  The function calls eq_location_clean
to clean the LOCATION_NAME column into a format suitable for display purposes.
}
\examples{
\dontrun{
filename<-system.file("extdata","data.gz",package="ykvdpcap")
readr::read_delim(filename, delim = "\\t") \%>\%
eq_clean_data()
}

}
