% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_functions.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Creates popup text for markers.}
\usage{
eq_create_label(eq_clean = NULL)
}
\arguments{
\item{eq_clean}{The clean earthquake data in a tbl_df object.}
}
\value{
This function returns a character vector containing
   popup text to be used in a leaflet visualization.
}
\description{
This function generates HTML formatted text to be used in
popups for map markers.
}
\examples{
\dontrun{
filename<-system.file("extdata","data.gz",package="ykvdpcap")

data_with_popup<-readr::read_delim(filename, delim = "\\t") \%>\%
eq_clean_data() \%>\%
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
dplyr::mutate(popup_text = eq_create_label(.))


}

}
