% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_functions.R
\name{eq_data_read}
\alias{eq_data_read}
\title{Read file}
\usage{
eq_data_read(filename)
}
\arguments{
\item{filename}{A character string which is the name of the file
with the NOAA earthquake date}
}
\value{
This function returns the earthquake data in a tbl_df object
}
\description{
This is a function which reads the NOAA earthquake data from a file
into a dataframe and converts it into a tbl_df object
}
\note{
If the filename does not exist, the function stops with an
   error message.
}
\examples{
\dontrun{
filename<-system.file("extdata","data.gz",package="ykvdpcap")
eq_data_read(filename)
}

}
