% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_functions.R
\name{eq_map}
\alias{eq_map}
\title{Creates interactive map visualization of NOAA earthquake data.}
\usage{
eq_map(eq_clean = NULL, annot_col = "DATE")
}
\arguments{
\item{eq_clean}{The clean earthquake data in a tbl_df object.}

\item{annot_col}{Column in the tbl_df object to be used for annotation.}
}
\value{
This function returns an interactive map.
}
\description{
This function generates a map visualization for the NOAA earthquake
data.  The data should be subsetted in order to create an understandable
visualization.  The earthquakes are mapped based on the latitude and
longitude of the earthquake epicenter.  The function is a wrapper function
and calls the leaflet function to generate the map.  The epicenter is
annotated based on an annot_col which the user can specify.  In addition,
if the user specifies "popup_text" then a call to eq_create_label generates
the appropriate text for the popup.
}
\note{
If an invalid column name is provided, the function provides a warning
and uses the LOCATION_NAME column as teh annotation column.
}
\examples{
\dontrun{

filename<-system.file("extdata","data.gz",package="ykvdpcap")
readr::read_delim(filename, delim = "\\t") \%>\%
eq_clean_data() \%>\%
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
eq_map(annot_col = "DATE")


filename<-system.file("extdata","data.gz",package="ykvdpcap")
readr::read_delim(filename, delim = "\\t") \%>\%
eq_clean_data() \%>\%
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
eq_map(annot_col = "popup_text")

}

}
\references{
\url{http://rstudio.github.io/leaflet/}
}
