% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_functions.R
\name{eq_time}
\alias{eq_time}
\title{Creates a timeline visualization for NOAA earthquake data.}
\usage{
eq_time(eq_clean = NULL, y = NULL, size = NULL, color = NULL,
  alpha = 0.4, timeline_label = FALSE, n_max = 5)
}
\arguments{
\item{eq_clean}{The clean earthquake data in a tbl_df object.}

\item{y}{Optional parameter for COUNTRY; specifes a separate
timeline for each country.}

\item{size}{Optional parameter for earthquake magnitude (EQ_PRIMARY);
specifies a point size based on magnitude.}

\item{color}{Optional parameter for number of deaths (DEATHS);
specifies a point color based on deaths.}

\item{alpha}{Optional parameter for point transparency.}

\item{timeline_label}{Optional parameter to select timeline labeling.}

\item{n_max}{Optional parameter to specify number of earthquakes to
be labeled.  The earthquakes with n_max highest magnitudes will be
labeled.}
}
\value{
This function returns a timeline visualization for a subset
   of the earthquake data.
}
\description{
This function generates a timeline visualization for the NOAA
earthquake data.  The data should be subsetted in order to create
an understandable visualization.  There are several ways to call
the function so that different visualizations are generated.  The
function is a wrapper function and calls geom_timeline and if a
label is specified, also calls the geom_timeline_label function.
}
\examples{
\dontrun{

filename<-system.file("extdata","data.gz",package="ykvdpcap")

readr::read_delim(filename, delim = "\\t") \%>\%
eq_clean_data() \%>\%
dplyr::filter(COUNTRY == "USA" & lubridate::year(DATE) >= 2000) \%>\%
eq_time(size="EQ_PRIMARY",color="DEATHS")

readr::read_delim(filename, delim = "\\t") \%>\%
eq_clean_data() \%>\%
dplyr::filter(COUNTRY == "USA" & lubridate::year(DATE) >= 2000) \%>\%
eq_time(size="EQ_PRIMARY",color="DEATHS",alpha=0.5,timeline_label=TRUE,n_max=8)

readr::read_delim(filename, delim = "\\t") \%>\%
eq_clean_data() \%>\%
dplyr::filter((COUNTRY=="USA" | COUNTRY=="CHINA") & lubridate::year(DATE) >= 2000) \%>\%
eq_time(y="COUNTRY",color="DEATHS",alpha=0.5)

readr::read_delim(filename, delim = "\\t") \%>\%
eq_clean_data() \%>\%
dplyr::filter((COUNTRY=="USA" | COUNTRY=="CHINA") & lubridate::year(DATE) >= 2000) \%>\%
eq_time(y="COUNTRY",color="DEATHS",alpha=0.5,timeline_label=TRUE)

}
}
