% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_functions.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Create a timeline geom}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{A set of aesthetic mappings.  If null, uses the
default setting in ggplot()}

\item{data}{A dataset If null, uses the default data specified
in ggplot()}

\item{stat}{The name of the statistical transformation to use.
Identity means keep the data as is.}

\item{position}{The method used to adjust overlapping objects.
Identity means keep the position as is.}

\item{na.rm}{Flag to remove values with NA}

\item{show.legend}{Flag to say whether layer should be present in legend.}

\item{inherit.aes}{Flag to say whether default aesthetics should be inherited.}

\item{...}{Additional arguments passed to layer()}
}
\value{
This function returns a layer that can be added to a plot
  using the ggplot function
}
\description{
This function adds a geom which displays a timeline for
NOAA earthquake data.  This is the basic timeline visualization
without additional modifications.  For a cleaner visualization
the wrapper function, eq_time, should be used.
}
\note{
The parameters for GeomTimeLine include columns from the earthquake data:
  Required parameters: x=DATE
  Optional parameters: y=COUNTRY,size=EQ_PRIMARY,color=DEATHS,alpha=x where x
  is a number from 0 to 1
}
\examples{
\dontrun{
readr::read_delim(filename, delim = "\\t") \%>\%
eq_clean_data() \%>\%
dplyr::filter(COUNTRY == "USA" & lubridate::year(DATE) >= 2000) \%>\%
ggplot() +
geom_timeline(aes(x=DATE,size=EQ_PRIMARY,color=DEATHS),alpha=0.5)

readr::read_delim(filename, delim = "\\t") \%>\%
eq_clean_data() \%>\%
dplyr::filter((COUNTRY=="USA" | COUNTRY=="CHINA") & lubridate::year(DATE) >= 2000) \%>\%
ggplot() +
geom_timeline(aes(x=DATE,y=COUNTRY,size=EQ_PRIMARY,color=DEATHS),alpha=0.5)

}

}
